requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.6
django-crispy-forms==2.1
sqlparse==0.5.0
Pillow==10.3.0 
django-colorfield==0.11.0 
django-filter==24.2
mysqlclient==2.2.4
faker==25.8.0 
python-dateutil==2.9.0.post0 
six==1.16.0
python-dotenv==1.0.1
algoliasearch==3.0.0 
algoliasearch==django-3.0.0
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
requests==2.32.3 
urllib3==2.2.1